finished all tests.

+++ code.js
@@ -34,5 +34,5 @@ function add(num1, num2) {
     if (isNaN(num1) || isNaN(num2)) {
         return NaN;
     } else
-    return (parseFloat(num1) + parseFloat(num2));
+    return (parseInt(num1) + parseInt(num2));
 }
